requirements.txt
fastapi
uvicorn
python-multipart
pillow
pytesseract
transformers
torch
numpy
opencv-python